@@ -1,6 +1,6 @@
 {
   "name": "docz-core",
-  "version": "1.0.2",
+  "version": "1.2.0",
   "description": "All docz core logic of bundle and parsing is included on this package",
   "license": "MIT",
   "main": "dist/index.js",
@@ -14,87 +14,49 @@
   "scripts": {
     "dev": "cross-env NODE_ENV=development yarn build -w",
     "build": "cross-env NODE_ENV=production rollup -c",
-    "fix": "run-s fix:*",
-    "fix:prettier": "prettier \"src/**/*.{ts,tsx,md,mdx,js,jsx,json}\" --write",
-    "fix:tslint": "tslint --fix --project .",
-    "tslint": "tslint --project ."
+    "fix": "yarn lint --fix",
+    "lint": "eslint . --ext .ts,.tsx",
+    "precommit": "lint-staged"
   },
   "dependencies": {
-    "@babel/core": "7.4.3",
-    "@babel/polyfill": "7.4.3",
-    "@babel/runtime": "^7.4.3",
-    "@mdx-js/loader": "^1.0.10",
     "@sindresorhus/slugify": "^0.9.1",
-    "@svgr/webpack": "^4.2.0",
-    "acorn": "6.0.5",
-    "babel-loader": "^8.0.5",
-    "babel-plugin-export-metadata": "^1.0.2",
-    "babel-plugin-named-asset-import": "^0.3.1",
-    "babel-preset-react-app": "^7.0.2",
-    "cache-loader": "^2.0.1",
     "chalk": "^2.4.2",
-    "chokidar": "^2.1.5",
-    "common-tags": "^1.8.0",
+    "chokidar": "^3.0.2",
     "detect-port": "^1.3.0",
-    "docz-utils": "^1.0.2",
-    "dotenv": "^7.0.0",
+    "docz-utils": "^1.2.0",
     "env-dot-prop": "^2.0.1",
-    "express": "^4.16.4",
     "fast-deep-equal": "^2.0.1",
-    "fast-glob": "^2.2.6",
-    "file-loader": "^3.0.1",
-    "find-up": "^3.0.0",
-    "friendly-errors-webpack-plugin": "^1.7.0",
-    "fs-extra": "^7.0.1",
+    "fast-glob": "^3.0.4",
+    "find-up": "^4.1.0",
+    "fs-extra": "^8.1.0",
     "get-pkg-repo": "^4.1.0",
-    "html-minifier": "^4.0.0",
     "humanize-string": "^2.1.0",
-    "load-cfg": "^1.0.2",
-    "lodash": "^4.17.11",
-    "mini-html-webpack-plugin": "^0.2.3",
-    "p-reduce": "^2.1.0",
-    "react-dev-utils": "^8.0.0",
-    "react-docgen": "^4.1.0",
-    "react-docgen-actual-name-handler": "^1.0.2",
-    "react-docgen-external-proptypes-handler": "^1.0.2",
-    "react-docgen-typescript": "^1.12.3",
-    "react-hot-loader": "^4.8.4",
-    "recast": "^0.17.5",
-    "rehype-docz": "^1.0.2",
-    "rehype-slug": "^2.0.2",
-    "remark-docz": "^1.0.2",
-    "remark-frontmatter": "^1.3.1",
-    "remark-parse": "^6.0.2",
-    "resolve": "^1.10.0",
+    "latest-version": "^5.1.0",
+    "load-cfg": "^1.2.0",
+    "lodash": "^4.17.14",
+    "minimatch": "^3.0.4",
+    "ora": "^3.4.0",
+    "react-docgen": "^4.1.1",
+    "react-docgen-actual-name-handler": "^1.2.0",
+    "react-docgen-external-proptypes-handler": "^1.0.3",
+    "react-docgen-typescript": "^1.12.5",
+    "recast": "^0.18.1",
+    "resolve": "^1.11.1",
+    "shelljs": "^0.8.3",
     "signale": "^1.4.0",
-    "source-map-loader": "^0.2.4",
-    "terser-webpack-plugin": "^1.2.3",
-    "thread-loader": "^2.1.2",
     "titleize": "^2.1.0",
-    "typescript": "3.3.4000",
-    "url-loader": "^1.1.2",
-    "webpack": "^4.30.0",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-chain": "^5.2.4",
-    "webpack-dev-server": "^3.3.1",
-    "webpack-hot-client": "^4.1.1",
-    "webpack-manifest-plugin": "^2.0.4",
-    "webpackbar": "^3.1.5",
-    "ws": "^6.2.1",
-    "yargs": "^13.2.2"
+    "typescript": "3.5.3",
+    "xstate": "^4.6.7",
+    "yargs": "^13.2.4"
   },
   "devDependencies": {
-    "@types/express": "^4.16.1",
+    "@types/cross-spawn": "^6.0.0",
     "@types/find-up": "^2.1.1",
-    "@types/html-minifier": "^3.5.3",
-    "@types/p-reduce": "^1.0.1",
-    "@types/react": "^16.8.13",
     "@types/resolve": "^0.0.8",
-    "@types/webpack": "^4.4.27",
-    "@types/webpack-chain": "^5.2.0",
-    "@types/webpack-dev-server": "^3.1.5",
-    "@types/ws": "^6.0.1",
-    "cpy": "^7.2.0",
-    "docz-rollup": "^1.0.2"
+    "@types/shelljs": "^0.8.5",
+    "@types/signale": "^1.2.1",
+    "cpy": "^7.3.0",
+    "docz-rollup": "^1.2.0",
+    "eslint-config-docz-ts": "^1.2.0"
   }
 }
