@@ -1,10 +1,10 @@
-import * as path from 'path'
 import chokidar from 'chokidar'
 import equal from 'fast-deep-equal'
 import { get } from 'lodash/fp'
 
+import { WATCH_IGNORE } from './config'
 import { Params, State } from '../lib/DataServer'
-import { Entries } from '../lib/Entries'
+import { Entries, getFilesToMatch } from '../lib/Entries'
 import { Config } from '../config/argv'
 import * as paths from '../config/paths'
 
@@ -18,22 +18,20 @@ const updateEntries = (entries: Entries) => async (p: Params) => {
   const map = await entries.get()
 
   if (map && !equal(prev, map)) {
-    await Entries.writeImports(map)
     p.setState('entries', mapToArray(map))
   }
 }
 
-export const state = (entries: Entries, config: Config): State => {
-  const src = path.relative(paths.root, config.src)
-  const files = Array.isArray(config.files)
-    ? config.files.map(filePath => path.join(src, filePath))
-    : path.join(src, config.files)
-
-  const ignored = config.watchIgnore || /(((^|[\/\\])\..+)|(node_modules))/
-  const watcher = chokidar.watch(files, {
-    cwd: paths.root,
+export const state = (
+  entries: Entries,
+  config: Config,
+  dev?: boolean
+): State => {
+  const ignored = config.watchIgnore || WATCH_IGNORE
+  const watcher = chokidar.watch(getFilesToMatch(config), {
     ignored,
     persistent: true,
+    cwd: paths.getRootDir(config),
   })
 
   watcher.setMaxListeners(Infinity)
@@ -42,16 +40,18 @@ export const state = (entries: Entries, config: Config): State => {
     id: 'entries',
     start: async params => {
       const update = updateEntries(entries)
-
       await update(params)
-      watcher.on('add', async () => update(params))
-      watcher.on('change', async () => update(params))
-      watcher.on('unlink', async () => update(params))
-      watcher.on('raw', async (event: string, path: string, details: any) => {
-        if (details.event === 'moved' && details.type === 'directory') {
-          await update(params)
-        }
-      })
+
+      if (dev) {
+        watcher.on('add', async () => update(params))
+        watcher.on('change', async () => update(params))
+        watcher.on('unlink', async () => update(params))
+        watcher.on('raw', async (event: string, path: string, details: any) => {
+          if (details.event === 'moved' && details.type === 'directory') {
+            await update(params)
+          }
+        })
+      }
     },
     close: () => {
       watcher.close()
