@@ -1,10 +1,12 @@
 import * as fs from 'fs-extra'
+import * as path from 'path'
 import logger from 'signale'
-import externalProptypesHandler from 'react-docgen-external-proptypes-handler'
+import externalProptypesHandler from './externalProptypesHandler'
 import actualNameHandler from 'react-docgen-actual-name-handler'
 import reactDocgen from 'react-docgen'
 
 import { Config } from '../../config/argv'
+import { getRootDir } from '../../config/paths'
 
 const throwError = (err: any) => {
   logger.fatal(`Error parsing static types`)
@@ -16,16 +18,18 @@ export const jsParser = (files: string[], config: Config) => {
     config.docgenConfig.resolver ||
     reactDocgen.resolver.findAllExportedComponentDefinitions
 
+  const root = getRootDir(config)
   const parseFilepathProps = (filepath: string) => {
+    const fullpath = path.resolve(root, filepath)
     const handlers = reactDocgen.defaultHandlers.concat([
       externalProptypesHandler(filepath),
       actualNameHandler,
     ])
 
     try {
-      const code = fs.readFileSync(filepath, 'utf-8')
+      const code = fs.readFileSync(fullpath, { encoding: 'utf-8' })
       const props = reactDocgen.parse(code, resolver, handlers)
-      return { key: filepath, value: props }
+      return { key: path.normalize(filepath), value: props }
     } catch (err) {
       if (config.debug) throwError(err)
       return null
