@@ -1,13 +1,8 @@
 import * as React from 'react'
 import { useContext, createContext } from 'react'
 import { Fragment, SFC, ComponentType as CT } from 'react'
-import { RouteComponentProps } from '@reach/router'
 
 import { Entry } from '../state'
-
-export type PageProps = RouteComponentProps<any> & {
-  doc: Entry
-}
 
 export interface PlaygroundProps {
   className?: string
@@ -17,39 +12,24 @@ export interface PlaygroundProps {
   component: JSX.Element
   position: number
   code: string
-  codesandbox: string
   scope: Record<string, any>
 }
 
-export type PlaygroundComponent = CT<PlaygroundProps>
-
-export interface ComponentsMap {
-  loading?: CT
-  page?: CT<PageProps>
-  notFound?: CT<RouteComponentProps<any>>
-  playground?: PlaygroundComponent
-  h1?: CT<any> | string
-  h2?: CT<any> | string
-  h3?: CT<any> | string
-  h4?: CT<any> | string
-  h5?: CT<any> | string
-  h6?: CT<any> | string
-  span?: CT<any> | string
-  a?: CT<any> | string
-  ul?: CT<any> | string
-  table?: CT<any> | string
-  pre?: CT<any> | string
-  code?: CT<any> | string
-  inlineCode?: CT<any> | string
+export interface LayoutProps {
+  doc: Entry
   [key: string]: any
 }
 
-export type NotFoundComponent = CT<RouteComponentProps<any>>
+export interface ComponentsMap {
+  notFound?: CT
+  layout?: CT<LayoutProps>
+  playground?: CT<PlaygroundProps>
+  [key: string]: any
+}
 
-const DefaultNotFound: NotFoundComponent = () => <Fragment>Not found</Fragment>
-const DefaultLoading: SFC = () => <Fragment>Loading</Fragment>
-const DefaultPage: SFC<any> = ({ children }) => <Fragment>{children}</Fragment>
-const DefaultPlayground: PlaygroundComponent = ({ component, code }) => (
+const DefNotFound: SFC = () => <Fragment>Not found</Fragment>
+const DefLayout: SFC = ({ children }) => <Fragment>{children}</Fragment>
+const DefPlayground: SFC<PlaygroundProps> = ({ component, code }) => (
   <Fragment>
     {component}
     {code}
@@ -57,18 +37,16 @@ const DefaultPlayground: PlaygroundComponent = ({ component, code }) => (
 )
 
 const defaultComponents: ComponentsMap = {
-  loading: DefaultLoading,
-  playground: DefaultPlayground,
-  notFound: DefaultNotFound,
-  page: DefaultPage,
+  layout: DefLayout,
+  notFound: DefNotFound,
+  playground: DefPlayground,
 }
-
-const ctx = createContext<ComponentsMap>({})
 
 export interface ComponentsProviderProps {
   components: ComponentsMap
 }
 
+const ctx = createContext<ComponentsMap>({})
 export const ComponentsProvider: SFC<ComponentsProviderProps> = ({
   components: themeComponents = {},
   children,
