@@ -1,4 +1,7 @@
-exports.createPages = require('./src/node/createPages')
-exports.onCreateBabelConfig = require('./src/node/onCreateBabelConfig')
-exports.onPreInit = require('./src/node/onPreInit')
-exports.sourceNodes = require('./src/node/sourceNodes')
+exports.createPages = require('./lib/createPages')
+exports.onCreateBabelConfig = require('./lib/onCreateBabelConfig')
+exports.onCreateDevServer = require('./lib/onCreateDevServer')
+exports.onCreateWebpackConfig = require('./lib/onCreateWebpackConfig')
+exports.onPreBuild = require('./lib/sourceNodes')
+exports.onPostBuild = require('./lib/sourceNodes')
+exports.sourceNodes = require('./lib/sourceNodes')
