@@ -1,7 +1,7 @@
 // @flow
 
 import React, { Fragment } from 'react'
-import styled from 'styled-components'
+import styled from '@emotion/styled'
 
 const kinds = {
   info: '#5352ED',
