@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react'
 import { Text as BaseText } from 'react-native'
-import styled from 'styled-components/native'
+import styled from '@emotion/primitives'
 import * as t from 'prop-types'
 
 const kinds = {
