@@ -1,5 +1,5 @@
 export default {
-  title: 'Docz Typescript',
-  codeSandbox: false,
   typescript: true,
+  title: 'Docz Typescript',
+  menu: ['Getting Started', 'Components'],
 }
