@@ -1,5 +1,5 @@
 import React, { SFC } from 'react'
-import styled from 'styled-components'
+import styled from '@emotion/styled'
 
 export type Kind = 'info' | 'positive' | 'negative' | 'warning'
 export type KindMap = Record<Kind, string>
@@ -19,12 +19,12 @@ export interface AlertProps {
   kind: 'info' | 'positive' | 'negative' | 'warning'
 }
 
-const AlertStyled = styled('div')`
+const AlertStyled = styled('div')<AlertProps>`
   padding: 15px 20px;
   background: white;
   border-radius: 3px;
   color: white;
-  background: ${({ kind = 'info' }: AlertProps) => kinds[kind]};
+  background: ${({ kind = 'info' }) => kinds[kind]};
 `
 
 export const Alert: SFC<AlertProps> = ({ kind, ...props }) => (
