@@ -5,7 +5,6 @@ const { get } = require('lodash')
 const buildFileMeta = template(`
   if (typeof ID !== 'undefined' && ID && ID === Object(ID) && Object.isExtensible(ID)) {
     Object.defineProperty(ID, '__filemeta', {
-      enumerable: true,
       configurable: true,
       value: {
         name: NAME,
@@ -21,11 +20,10 @@ const replaceExportDefault = template(`
 `)
 
 const getFilename = state => {
+  const rootDir = get(state, 'opts.root', process.cwd())
   const filename = get(state, 'file.opts.filename')
   let filePath = filename && path.relative(process.cwd(), filename)
-  if (process.platform === 'win32') {
-    filePath = filePath.split('\\').join('/')
-  }
+  if (process.platform === 'win32') filePath = filePath.split('\\').join('/')
   return filePath
 }
 
@@ -67,13 +65,6 @@ const renameDefaultAddFileMetaProperties = (t, path, filename, name) => {
     SOURCE: sourceValue,
   })
 
-  // insert
-  const newNode = buildFileMeta({
-    ID: t.identifier(fallbackName),
-    NAME: t.stringLiteral(fallbackName),
-    FILENAME: t.stringLiteral(filename),
-  })
-
   pathToInsert.replaceWithMultiple(nameExport)
 }
 
@@ -92,7 +83,7 @@ const insertNodeExport = t => (path, state) => {
       const declarationName = get(declaration, 'id.name')
       addFileMetaProperties(t, path, filename, declarationName)
     }
-  } else if (specifiers) {
+  } else if (specifiers && !state.opts.notUseSpecifiers) {
     for (specifier of specifiers) {
       const localName = get(specifier, 'local.name')
       const exportedName = get(specifier, 'exported.name')
